python/requirements.txt
requests>=2.28.1
pandas>=1.5.0
numpy>=1.23.3
prometheus-client>=0.16.0
psycopg2-binary>=2.9.3
python-dotenv>=0.21.0
pyyaml>=6.0
pyracing>=0.1.4
matplotlib>=3.6.0
pytest>=7.0.0
black>=22.8.0
flake8>=5.0.4
jsonschema>=4.17.0
# irsdk>=1.3.0 - Install separately from GitHub
